% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{filtrer_par_genre}
\alias{filtrer_par_genre}
\title{Filtrer les données par genre}
\usage{
filtrer_par_genre(data, g)
}
\arguments{
\item{data}{Le dataframe marathon2}

\item{g}{Le genre à filtrer ("Male" ou "Female")}
}
\value{
Un dataframe filtre
}
\description{
Filtrer les données par genre
}
