% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marathon2}
\alias{marathon2}
\title{Résultats des marathons mondiaux}
\format{
\subsection{\code{marathon2}}{

A data frame with 536 rows and 7 columns:
\describe{
\item{year}{Année de la course}
\item{winner}{Nom du vainqueur}
\item{gender}{Sexe (Male/Female)}
\item{country}{Pays d'origine}
\item{time}{Temps au format H:M:S}
\item{marathon}{Nom de la ville du marathon}
\item{time_seconds}{Temps total en secondes}
}
}
}
\source{
https://www.kaggle.com/datasets/emmanuelleai/world-marathons-majors
}
\usage{
marathon2
}
\description{
Un jeu de données contenant tous les vainqueurs des marathons majors.
}
\keyword{datasets}
