% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{plot_marathon_box}
\alias{plot_marathon_box}
\title{Graphique de repartition des temps par genre (Boxplot)}
\usage{
plot_marathon_box(data)
}
\arguments{
\item{data}{Le dataframe marathon2}
}
\description{
Graphique de repartition des temps par genre (Boxplot)
}
