% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{records_par_pays}
\alias{records_par_pays}
\title{Trouver les records par pays}
\usage{
records_par_pays(data)
}
\arguments{
\item{data}{Le dataframe marathon2}
}
\description{
Trouver les records par pays
}
